+++ src/containerapp/azext_containerapp/_utils.py
@@ -726,3 +726,11 @@ def parse_build_env_vars(env_list):
         })
 
     return env_var_def
+
+
+def is_cloud_supported_by_connected_env(cli_ctx):
+    from azure.cli.core.cloud import CloudNameEnum
+    if cli_ctx.cloud.name == CloudNameEnum.AzureCloud:
+        return True
+
+    return False